Move Condition to the foundation module.
